pyproject.toml
[tool.poetry]
name = "py-game"
version = "0.1.0"
description = ""
authors = ["Lucas Garcia <lsglucas@protonmail.com>"]
readme = "README.md"
packages = [{include = "py_game"}]

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.3.2"
pytmx = "^3.32"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
